+++ back-end/plogging/src/main/java/com/backend/plogging/dto/response/plogging/RecordResponseDto.java
@@ -20,10 +20,10 @@ public class RecordResponseDto {
 
     private Long recordId;
     private Float distance;
-    private Float startLat;
-    private Float startLng;
-    private Float endLat;
-    private Float endLng;
+    private Double startLat;
+    private Double startLng;
+    private Double endLat;
+    private Double endLng;
     private Float runningTime;
     private Integer kcal;
